+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to **ValueStringBuilder** will be documented in this file. T
 
 This is the `v3` major release. The API is almost the same as in `v2` - there is only a slight change in the `Concat` static helper method to reflect a less-boxed API.
 
+### Added
+- .NET 10.0 support
+
 ### Changed
 - `ValueStringBuilder.Concat` uses `params ReadOnlySpan<T>` to reduce boxing and improve performance.